refactor(core): extract transcript projection
