+++ packages/core/src/session/projection/transcript.ts
@@ -1,0 +1,164 @@
+export * as TranscriptProjection from "./transcript"
+
+import { and, desc, eq, sql } from "drizzle-orm"
+import { DateTime, Effect, Schema } from "effect"
+import { Database } from "../../database/database"
+import { ModelV2 } from "../../model"
+import { SessionEvent } from "../event"
+import { SessionMessage } from "../message"
+import { SessionMessageUpdater } from "../message-updater"
+import { SessionMessageTable, SessionTable } from "../sql"
+
+type DatabaseService = Database.Interface["db"]
+type CurrentDurableEvent = Extract<SessionEvent.Event, { readonly durable: object }>
+type MessageEvent = Exclude<
+  CurrentDurableEvent,
+  typeof SessionEvent.Forked.Type | typeof SessionEvent.Deleted.Type | typeof SessionEvent.InstructionsUpdated.Type
+>
+
+const decodeMessage = Schema.decodeUnknownSync(SessionMessage.Info)
+const encodeMessage = Schema.encodeSync(SessionMessage.Info)
+
+export function project(db: DatabaseService, event: MessageEvent) {
+  return Effect.gen(function* () {
+    const decodeRow = (row: typeof SessionMessageTable.$inferSelect) =>
+      decodeMessage({ ...row.data, id: row.id, type: row.type })
+    const updateMessage = (message: SessionMessage.Info) => {
+      if (event.durable === undefined)
+        return Effect.die(new Error("Durable Session event is missing aggregate sequence"))
+      const encoded = encodeMessage(message)
+      const { id, type, ...data } = encoded
+      return db
+        .update(SessionMessageTable)
+        .set({ type, time_created: DateTime.toEpochMillis(message.time.created), data })
+        .where(
+          and(
+            eq(SessionMessageTable.id, SessionMessage.ID.make(id)),
+            eq(SessionMessageTable.session_id, event.data.sessionID),
+          ),
+        )
+        .run()
+        .pipe(Effect.orDie)
+    }
+    const appendMessage = (message: SessionMessage.Info) => appendAtEventSequence(db, event, message)
+    const adapter: SessionMessageUpdater.Adapter = {
+      getModel() {
+        return db
+          .select({ model: SessionTable.model })
+          .from(SessionTable)
+          .where(eq(SessionTable.id, event.data.sessionID))
+          .get()
+          .pipe(
+            Effect.orDie,
+            Effect.map((row) => (row?.model ? Schema.decodeUnknownSync(ModelV2.Ref)(row.model) : undefined)),
+          )
+      },
+      getCurrentAssistant() {
+        return Effect.gen(function* () {
+          // A newer step supersedes stale incomplete rows; never resume an older assistant projection.
+          const row = yield* db
+            .select()
+            .from(SessionMessageTable)
+            .where(
+              and(eq(SessionMessageTable.session_id, event.data.sessionID), eq(SessionMessageTable.type, "assistant")),
+            )
+            .orderBy(desc(SessionMessageTable.seq))
+            .limit(1)
+            .get()
+            .pipe(Effect.orDie)
+          if (!row) return undefined
+          const message = decodeRow(row)
+          return message.type === "assistant" && !message.time.completed ? message : undefined
+        })
+      },
+      getAssistant(messageID) {
+        return Effect.gen(function* () {
+          const row = yield* db
+            .select()
+            .from(SessionMessageTable)
+            .where(
+              and(
+                eq(SessionMessageTable.id, messageID),
+                eq(SessionMessageTable.session_id, event.data.sessionID),
+                eq(SessionMessageTable.type, "assistant"),
+              ),
+            )
+            .get()
+            .pipe(Effect.orDie)
+          if (!row) return undefined
+          const message = decodeRow(row)
+          return message.type === "assistant" ? message : undefined
+        })
+      },
+      getShell(shellID) {
+        return Effect.gen(function* () {
+          const row = yield* db
+            .select()
+            .from(SessionMessageTable)
+            .where(
+              and(
+                eq(SessionMessageTable.session_id, event.data.sessionID),
+                eq(SessionMessageTable.type, "shell"),
+                sql`json_extract(${SessionMessageTable.data}, '$.shellID') = ${shellID}`,
+              ),
+            )
+            .orderBy(desc(SessionMessageTable.seq))
+            .limit(1)
+            .get()
+            .pipe(Effect.orDie)
+          if (!row) return undefined
+          const message = decodeRow(row)
+          return message.type === "shell" ? message : undefined
+        })
+      },
+      getCompaction() {
+        return Effect.gen(function* () {
+          const row = yield* db
+            .select()
+            .from(SessionMessageTable)
+            .where(
+              and(
+                eq(SessionMessageTable.session_id, event.data.sessionID),
+                eq(SessionMessageTable.type, "compaction"),
+                sql`json_extract(${SessionMessageTable.data}, '$.status') = 'running'`,
+              ),
+            )
+            .orderBy(desc(SessionMessageTable.seq))
+            .limit(1)
+            .get()
+            .pipe(Effect.orDie)
+          if (!row) return undefined
+          const message = decodeRow(row)
+          return message.type === "compaction" ? message : undefined
+        })
+      },
+      updateAssistant: updateMessage,
+      updateShell: updateMessage,
+      updateCompaction: updateMessage,
+      appendMessage,
+    }
+    yield* SessionMessageUpdater.update(adapter, event)
+  })
+}
+
+export function appendAtEventSequence(
+  db: DatabaseService,
+  event: SessionEvent.DurableEvent,
+  message: SessionMessage.Info,
+) {
+  if (event.durable === undefined) return Effect.die(new Error("Durable Session event is missing aggregate sequence"))
+  const encoded = encodeMessage(message)
+  const { id, type, ...data } = encoded
+  return db
+    .insert(SessionMessageTable)
+    .values({
+      id: SessionMessage.ID.make(id),
+      session_id: event.data.sessionID,
+      type,
+      seq: event.durable.seq,
+      time_created: DateTime.toEpochMillis(message.time.created),
+      data,
+    })
+    .run()
+    .pipe(Effect.orDie)
+}
